# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance concurrent data ingestion and processing pipeline.",
        "title": "Data Processor Pipeline API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/pipelines": {
            "get": {
                "description": "Retrieves a list of all historical and running jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "List all pipeline jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Ingests data from the specified source URL and begins concurrent processing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Start a new pipeline job",
                "parameters": [
                    {
                        "description": "Pipeline Configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}": {
            "get": {
                "description": "Fetches the metadata and status of a specific job by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Get pipeline job details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (e.g. 1234-abcd)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a job and all associated artifacts from PostgreSQL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Delete job and artifacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}/cancel": {
            "patch": {
                "description": "Safely cancels an active pipeline via context cancellation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Cancel running pipeline job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}/errors": {
            "get": {
                "description": "Retrieves any failed records and error logs for a job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Get job error logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}/progress": {
            "get": {
                "description": "Retrieves processing metrics (total vs processed records) for a running job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Get real-time job progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pipelines/{id}/results": {
            "get": {
                "description": "Retrieves the aggregated mathematical outputs for a completed job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Get final job results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MockResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CreateJobRequest": {
            "type": "object",
            "properties": {
                "source_type": {
                    "type": "string",
                    "enum": [
                        "csv",
                        "json",
                        "mixed"
                    ],
                    "example": "csv"
                },
                "source_url": {
                    "type": "string",
                    "example": "https://covid.ourworldindata.org/data/owid-covid-data.csv"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "job not found"
                }
            }
        },
        "controllers.MockResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Endpoint hit successfully"
                }
            }
        }
    }
}